apps/api/requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Database
sqlalchemy>=2.0.30
asyncpg>=0.29.0
greenlet>=3.0.0
psycopg2-binary>=2.9.10
alembic>=1.13.0
pgvector>=0.3.0

# Storage
boto3>=1.34.0
supabase>=2.5.0

# AI providers
openai>=1.35.0
google-generativeai>=0.8.0

# Document parsing
pymupdf>=1.25.0
lxml>=5.2.0
python-multipart>=0.0.9

# OCR (optional)
pytesseract>=0.3.10
Pillow>=11.0.0

# Utilities
httpx>=0.27.0
python-dotenv>=1.0.0
